+++ packages/localizations/src/hu-HU.ts
@@ -612,6 +612,12 @@ export const huHU: LocalizationResource = {
       subtitle: 'hogy folytathasd a(z) {{applicationName}}',
       title: 'Ellenőrizd az emailed',
     },
+    emailCodeMfa: {
+      formTitle: 'Ellenőrizd az emailed',
+      resendButton: 'Nem kaptad meg a kódot? Újraküldés',
+      subtitle: 'hogy folytathasd a(z) {{applicationName}}',
+      title: 'Ellenőrizd az emailed',
+    },
     emailLink: {
       clientMismatch: {
         subtitle: undefined,
@@ -647,6 +653,12 @@ export const huHU: LocalizationResource = {
         titleNewTab: 'Egy másik lapon bejelezkeztél be',
       },
     },
+    emailLinkMfa: {
+      formSubtitle: 'Használja az e-mail címére küldött megerősítő linket',
+      resendButton: 'Nem kapta meg a linket? Újraküldés',
+      subtitle: 'a folytatáshoz a {{applicationName}} alkalmazásban',
+      title: 'Ellenőrizze az e-mailjét',
+    },
     enterpriseConnections: {
       subtitle: undefined,
       title: undefined,
@@ -664,6 +676,8 @@ export const huHU: LocalizationResource = {
       label__alternativeMethods: 'Vagy jelentkezz be más módon',
       title: 'Efelejtetted a jelszavad?',
     },
+    newDeviceVerificationNotice:
+      'Új eszközről jelentkezel be. Kérünk egy ellenőrzést, hogy fiókod biztonságban maradjon.',
     noAvailableMethods: {
       message: 'Nem lehet bejelentkezni. Nincs elérhető hitelesítő tényező.',
       subtitle: 'Hiba történt',